+++ src/main.js
@@ -2,10 +2,9 @@ import {MOCK_FILMS_COUNT, MOCK_COMMENTS_COUNT} from "./consts";
 import {render, RenderPosition} from "./utils/common";
 import {generateFilm} from "./mock/film";
 import {generateComment} from "./mock/comment";
-import Filter from "./view/filter";
 import Profile from "./view/profile";
 import FooterStatistics from "./view/footer";
-import FilmListsPesenter from "./presenter/film-lists";
+import FilmListsPresenter from "./presenter/film-lists";
 import FilterPresentor from "./presenter/filter";
 import FilmsModel from "./model/films";
 import CommentsModel from "./model/comments";
@@ -14,17 +13,10 @@ import FilterModel from "./model/filter";
 const films = new Array(MOCK_FILMS_COUNT).fill().map(generateFilm);
 const comments = new Array(MOCK_COMMENTS_COUNT).fill().map(generateComment);
 
+const filterModel = new FilterModel();
+
 const filmsModel = new FilmsModel();
 filmsModel.setFilms(films);
-
-const filterModel = new FilterModel();
-const filters = [
-  {
-    type: `all`,
-    name: `ALL`,
-    count: 0
-  }
-];
 
 const commentsModel = new CommentsModel();
 commentsModel.setComments(comments);
@@ -35,7 +27,7 @@ const footerStatisticsContainer = document.querySelector(`.footer`);
 
 const profile = new Profile(films);
 const footerStatistics = new FooterStatistics(films);
-const filmListsPresentor = new FilmListsPesenter(mainContainer, filmsModel, commentsModel, filterModel);
+const filmListsPresentor = new FilmListsPresenter(mainContainer, filmsModel, commentsModel, filterModel);
 const filterPresenter = new FilterPresentor(mainContainer, filterModel, filmsModel);
